# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vaara HTTP API
  version: "1.0.0"
  summary: Conformal-scoring risk evaluation and hash-chained audit emission for autonomous actions.
  description: |
    The Vaara HTTP API exposes the Vaara runtime kernel as a network surface.
    Any agent runtime, control plane, or evaluation framework can call the
    `/v1/score` endpoint to obtain a calibrated risk assessment for an action,
    and `/v1/audit/*` endpoints to append events to a hash-chained audit trail
    aligned with EU AI Act Articles 12 (record-keeping), 14 (human oversight),
    and 17 (quality management).

    The interface is implementation-agnostic. A reference server ships with
    `pip install vaara[server]`, but the spec is authoritative; alternative
    servers in any language can implement it without changing the contract.

    Versioning: paths are prefixed with `/v1/`. Breaking changes increment
    the path version. Additive changes do not.

    Authentication: when the server is started with an API key
    (`--api-key` / `VAARA_SERVER_API_KEY`), every endpoint except
    `GET /v1/health` requires `Authorization: Bearer <key>`; unauthenticated
    requests receive 401 with `error.code = "unauthenticated"`. The
    reference server refuses a non-loopback bind without a key, because an
    unauthenticated reachable instance allows policy hot-swap to
    allow-all, forged audit-event append, and calibration poisoning.
  license:
    name: AGPL-3.0-or-later
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
  /v1/score:
    post:
      tags: [score]
      summary: Score an action.
      description: |
        Submit an action context and receive a calibrated risk assessment with
        an `allow / escalate / deny` decision. The scorer is stateful: it
        maintains conformal calibration, MWU expert weights, and per-agent
        sequence-pattern history across requests.
      operationId: score
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScoreRequest" }
      responses:
        "200":
          description: Risk assessment.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScoreResponse" }
        "400":
          description: Malformed request.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Scorer unavailable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/score/outcome:
    post:
      tags: [score]
      summary: Report an outcome for online learning.
      description: |
        Report the observed outcome of a previously-scored action. Outcomes
        feed the MWU expert system and conformal calibration. An action's
        outcome can be reported once; subsequent reports for the same
        `action_id` are no-ops.
      operationId: reportOutcome
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OutcomeRequest" }
      responses:
        "204": { description: Outcome recorded. }
        "400":
          description: Malformed request.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Unknown action_id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/audit/events:
    post:
      tags: [audit]
      summary: Append an event to the audit chain.
      description: |
        Append a typed event to the hash-chained audit trail. The server
        returns the assigned `event_id` and `chain_position`. The event is
        cryptographically linked to the previous event; tampering invalidates
        the chain past the modified point.
      operationId: appendAuditEvent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuditEventRequest" }
      responses:
        "201":
          description: Event appended.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditEventResponse" }
        "400":
          description: Malformed event.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/audit/actions/{action_id}/chain:
    get:
      tags: [audit]
      summary: Read the audit chain for an action.
      description: |
        Return every event recorded for a given `action_id`, in order. Each
        event carries its hash linkage so an external auditor can verify
        integrity offline.
      operationId: readActionChain
      parameters:
        - name: action_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Chain segment for this action.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditChain" }
        "404":
          description: Unknown action_id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/audit/verify:
    post:
      tags: [audit]
      summary: Verify the audit chain.
      description: |
        Verify hash-chain integrity over a range of events. If `from_event_id`
        and `to_event_id` are omitted, the full chain is verified. Returns the
        first broken link if any, or a clean verification report.
      operationId: verifyAuditChain
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyRequest" }
      responses:
        "200":
          description: Verification result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerifyResponse" }
  /v1/detect/injection:
    post:
      tags: [detect]
      summary: Score text for prompt-injection likelihood.
      description: |
        Routes through the Vaara adversarial scorer (the same model
        behind vaara-bench-v1). Falls back to a small regex heuristic
        set when the `ml` extra is not installed; the backend field
        in the response reports which path served the call.
      operationId: detectInjection
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DetectInjectionRequest" }
      responses:
        "200":
          description: Detection result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetectInjectionResponse" }
  /v1/detect/pii:
    post:
      tags: [detect]
      summary: Scan text for PII.
      description: |
        Regex-based extractor over six categories: email, phone, ssn,
        ipv4, credit_card (Luhn-checked), iban (checksum-checked).
        Returns per-finding category, value, offset, and length so
        callers can highlight or redact in place.
      operationId: detectPII
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DetectPIIRequest" }
      responses:
        "200":
          description: Detection result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetectPIIResponse" }
  /v1/policy/reload:
    post:
      tags: [policy]
      summary: Atomically reload the running policy.
      description: |
        Parse and validate a new policy document, then swap it in atomically.
        The scorer's thresholds and sequence patterns rebind under its own
        lock so an evaluate() call in flight on another thread either sees
        the old pair or the new pair, never a torn half. Conformal
        calibration, MWU expert state, and agent profiles are preserved.

        Exactly one of `path` (server-side file path) or `body` (parsed
        policy document) must be supplied. The server must have been
        started with `vaara serve --policy PATH` for this endpoint to be
        available; otherwise it returns `409 policy_not_configured`. A
        malformed document returns `422 policy_invalid` and leaves the
        previously loaded policy in place.
      operationId: reloadPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PolicyReloadRequest" }
      responses:
        "200":
          description: Reload accepted, new policy active.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PolicyReloadResponse" }
        "400":
          description: Request supplied both `path` and `body`, or neither.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Server has no PolicyController; reload is disabled.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Policy document failed to parse or validate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key configured at server start; required on every endpoint except GET /v1/health when set.
  schemas:
    ScoreRequest:
      type: object
      required: [tool_name, agent_id]
      properties:
        tool_name:
          type: string
          maxLength: 512
          example: "tx.transfer"
        agent_id:
          type: string
          maxLength: 256
          example: "agent-007"
        action_type:
          type: string
          description: Optional category hint (transaction, data, infra, etc.).
        parameters:
          type: object
          description: Action parameters. Used for signal extraction.
          additionalProperties: true
        agent_confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
        base_risk_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        reversibility:
          type: string
          enum: [reversible, partially_reversible, irreversible]
        blast_radius:
          type: string
          enum: [local, account, organization, global]
        session_id:
          type: string
          maxLength: 256
        parent_action_id:
          type: string
          maxLength: 128
        context:
          type: object
          description: Arbitrary additional context. Not used by core scorer.
          additionalProperties: true
    ScoreResponse:
      type: object
      required: [action_id, decision, risk, thresholds, evaluation_ms]
      properties:
        action_id:
          type: string
          description: Server-assigned identifier. Use with /score/outcome and audit endpoints.
        decision:
          type: string
          enum: [allow, escalate, deny]
        risk:
          type: object
          required: [point, lower, upper, alpha]
          properties:
            point: { type: number, format: float, minimum: 0, maximum: 1 }
            lower: { type: number, format: float, minimum: 0, maximum: 1 }
            upper:
              type: number
              format: float
              minimum: 0
              maximum: 1
              description: Upper bound. The decision uses this conservatively.
            alpha:
              type: number
              format: float
              minimum: 0
              maximum: 1
              description: Effective miscoverage rate for this assessment.
            bucket:
              type: string
              nullable: true
              description: Mondrian bucket category, or null for marginal.
        signals:
          type: object
          description: Per-expert signal contributions, name to value in [0,1].
          additionalProperties: { type: number, format: float }
        mwu_weights:
          type: object
          description: Current MWU expert weights snapshot.
          additionalProperties: { type: number, format: float }
        thresholds:
          type: object
          required: [allow, deny]
          properties:
            allow: { type: number, format: float }
            deny: { type: number, format: float }
        sequence_risk: { type: number, format: float }
        calibration_size: { type: integer, minimum: 0 }
        evaluation_ms: { type: number, format: float }
        explanation: { type: string }
    OutcomeRequest:
      type: object
      required: [action_id, outcome_severity]
      properties:
        action_id: { type: string }
        outcome_severity:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: 0 = benign, 1 = catastrophic. Drives MWU weight update.
        notes: { type: string }
    AuditEventRequest:
      type: object
      required: [event_type, action_id]
      properties:
        event_type:
          type: string
          enum:
            - action_requested
            - risk_scored
            - decision_made
            - action_executed
            - action_blocked
            - escalation_sent
            - escalation_resolved
            - outcome_recorded
            - policy_override
        action_id: { type: string }
        agent_id: { type: string }
        tool_name: { type: string }
        payload:
          type: object
          description: Event-type-specific payload.
          additionalProperties: true
    AuditEventResponse:
      type: object
      required: [event_id, chain_position, event_hash, previous_hash, timestamp]
      properties:
        event_id: { type: string }
        chain_position: { type: integer, minimum: 0 }
        event_hash:
          type: string
          description: SHA-256 of this event including previous_hash.
          pattern: "^[0-9a-f]{64}$"
        previous_hash:
          type: string
          pattern: "^[0-9a-f]{64}$"
        timestamp:
          type: string
          format: date-time
    AuditChain:
      type: object
      required: [action_id, events]
      properties:
        action_id: { type: string }
        events:
          type: array
          items:
            type: object
            required: [event_id, event_type, chain_position, event_hash, previous_hash, timestamp]
            properties:
              event_id: { type: string }
              event_type: { type: string }
              chain_position: { type: integer }
              event_hash: { type: string }
              previous_hash: { type: string }
              timestamp: { type: string, format: date-time }
              payload:
                type: object
                additionalProperties: true
    VerifyRequest:
      type: object
      properties:
        from_event_id:
          type: string
          description: Inclusive start. Omit for chain start.
        to_event_id:
          type: string
          description: Inclusive end. Omit for chain end.
    VerifyResponse:
      type: object
      required: [valid, events_checked]
      properties:
        valid: { type: boolean }
        events_checked: { type: integer, minimum: 0 }
        first_break:
          type: object
          nullable: true
          description: Present iff valid is false. Identifies the first broken link.
          properties:
            event_id: { type: string }
            chain_position: { type: integer }
            expected_previous_hash: { type: string }
            actual_previous_hash: { type: string }
...
    DetectInjectionRequest:
      type: object
      required: [text]
      properties:
        text: { type: string, maxLength: 100000 }
        threshold:
          type: number
          format: float
          minimum: 0
          maximum: 1
          nullable: true
    DetectInjectionResponse:
      type: object
      required: [detected, score, threshold, bundle_version, backend]
      properties:
        detected: { type: boolean }
        score: { type: number, format: float, minimum: 0, maximum: 1 }
        threshold: { type: number, format: float, minimum: 0, maximum: 1 }
        bundle_version: { type: string }
        backend:
          type: string
          enum: [vaara_adversarial, heuristic]
    DetectPIIRequest:
      type: object
      required: [text]
      properties:
        text: { type: string, maxLength: 100000 }
    DetectPIIFinding:
      type: object
      required: [category, value, offset, length]
      properties:
        category:
          type: string
          enum: [email, phone, ssn, ipv4, credit_card, iban]
        value: { type: string }
        offset: { type: integer, minimum: 0 }
        length: { type: integer, minimum: 1 }
    DetectPIIResponse:
      type: object
      required: [detected, categories, findings]
      properties:
        detected: { type: boolean }
        categories:
          type: array
          items: { type: string }
        findings:
          type: array
          items: { $ref: "#/components/schemas/DetectPIIFinding" }
    PolicyReloadRequest:
      type: object
      description: |
        Exactly one of `path` or `body` must be supplied. `path` lets the
        server read the policy file directly; `body` carries an inline
        parsed document for cases where the operator runs on a different
        host than the server.
      properties:
        path: { type: string, maxLength: 4096 }
        body:
          type: object
          additionalProperties: true
        format:
          type: string
          enum: [json, yaml]
          nullable: true
    PolicyReloadResponse:
      type: object
      required:
        - version
        - thresholds_default
        - sequence_count
        - action_class_count
        - escalation_route_count
      properties:
        version:
          type: integer
          minimum: 1
          description: Monotonic generation counter; increments on every accepted reload.
        thresholds_default:
          type: object
          required: [escalate, deny]
          properties:
            escalate: { type: number, format: float, minimum: 0, maximum: 1 }
            deny: { type: number, format: float, minimum: 0, maximum: 1 }
        sequence_count: { type: integer, minimum: 0 }
        action_class_count: { type: integer, minimum: 0 }
        escalation_route_count: { type: integer, minimum: 0 }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true